+++ components/common/CustomCalendar.tsx
@@ -1,58 +1,28 @@
-import { BottomSheetMethods } from '@gorhom/bottom-sheet/lib/typescript/types';
-import { updateStartDateToDB } from 'api/supabaseAPI';
 import useKoreanToday from 'hooks/useKoreanToday';
-import { useState } from 'react';
-import { Text, TouchableHighlight, TouchableOpacity, View } from 'react-native';
+import { Text, View } from 'react-native';
 import { Calendar } from 'react-native-calendars';
-import { useIngredientStore } from 'stores/ingredientStore';
-import { useUserStore } from 'stores/userStore';
 import colors from 'tailwindcss/colors';
 import { getWeekAndDay } from 'utils/date';
 import { getWeekColor } from 'utils/getWeekColor';
 
 interface CustomCalendarProps {
-  bottomSheetRef?: React.RefObject<BottomSheetMethods>;
+  calendarKey?: React.Key;
+  selectedDate: string;
+  setSelectedDate: React.Dispatch<React.SetStateAction<string>>;
 }
 
-export default function CustomCalendar({ bottomSheetRef }: CustomCalendarProps) {
-  const startDate = useUserStore((state) => state.start_date);
+export default function CustomCalendar({
+  selectedDate,
+  setSelectedDate,
+  calendarKey,
+}: CustomCalendarProps) {
   const today = useKoreanToday();
 
-  const [selectedDate, setSelectedDate] = useState<string>(startDate ?? today);
-  const [calendarKey, setCalendarKey] = useState(false); // 캘린더 리렌더링용 키
   const { week, day } = getWeekAndDay(selectedDate);
-
-  const setDateToToday = () => {
-    setSelectedDate(today);
-    setCalendarKey((prev) => !prev);
-  };
-
-  const updateStartDate = async () => {
-    const setIngredients = useIngredientStore.getState().setIngredients;
-    const selectedIngredients = useIngredientStore.getState().selectedIngredients;
-
-    const filteredIngredients = selectedIngredients.filter((ingredient) => ingredient.week <= week);
-    setIngredients(filteredIngredients);
-    await updateStartDateToDB(selectedDate);
-
-    if (bottomSheetRef) {
-      bottomSheetRef.current.close();
-    }
-  };
   return (
     <View>
-      <View className="flex-row items-center justify-between">
-        <Text className="mb-2 text-2xl font-semibold">시작 날짜 재설정</Text>
-        <TouchableOpacity
-          className="rounded-lg border border-green-600 px-3 py-2.5"
-          onPress={setDateToToday}>
-          <Text className="text-green-600">오늘로 설정</Text>
-        </TouchableOpacity>
-      </View>
-      <Text className="mb-6 text-neutral-500">날짜를 선택해주세요</Text>
-
       <Calendar
-        key={String(calendarKey)}
+        key={calendarKey}
         current={selectedDate}
         maxDate={today}
         onDayPress={(day) => {
@@ -75,12 +45,6 @@ export default function CustomCalendar({ bottomSheetRef }: CustomCalendarProps) 
         <Text className={`text-lg ${getWeekColor(week)}`}>{week}주차</Text>
         <Text className="text-lg">{day}일부터 시작해요</Text>
       </View>
-
-      <TouchableHighlight onPress={updateStartDate}>
-        <View className="items-center rounded-lg bg-green-600 py-4">
-          <Text className="text-lg font-semibold text-white">날짜 재설정하기</Text>
-        </View>
-      </TouchableHighlight>
     </View>
   );
 }
